// src/components/cdk/cdk-ibor-preview/models/cdk-ibor-preview.model.ts
export interface ICDKIborPreviewData {
	id: string;
	partLabel: string;
	partValue: number;
	totalLabel: string;
	totalValue: number;
	showRatio: boolean;
	showCheckWheComplete: boolean;
	noDecimals: boolean;
	styling: object;
}

export interface ICDKIborPreviewLiterals {
	errorTitle: string;
	errorDesc: string;
}